requirements-prod.txt
Django==5.2
djangorestframework==3.16.0
gunicorn==23.0.0
celery>=5.5.3
redis>=6.2.0
drf-yasg==1.21.10
django-filter==25.1
django-import-export==4.3.7
django-jazzmin==3.0.1
django-simple-history>=3.0,<4.0
djangorestframework_simplejwt==5.5.0
django-extensions==4.1
django-cors-headers==4.7.0
django-redis==5.2.0
djoser==2.3.1
social-auth-app-django>=5.4.3
social-auth-core>=4.5.6
python-dotenv==1.1.0
requests==2.32.3
requests-oauthlib==2.0.0
PyJWT==2.9.0
Pillow==11.2.1
openpyxl==3.1.5
tablib==3.8.0
et_xmlfile==2.0.0
sqlparse==0.5.3
tzdata==2025.2
certifi==2025.1.31
urllib3==2.4.0
idna==3.10
charset-normalizer==3.4.1
